+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## master
+
+* Pod Repo Source Management, both in the Podfile Editor, and in the Main Menu  
+  [orta](https://github.com/orta)  
+  [#294](https://github.com/CocoaPods/CocoaPods-app/pull/294)
+
+
 ## [1.0.0.beta.8 2016.05.19](https://github.com/CocoaPods/CocoaPods-app/releases/tag/1.0.0.beta.8)
 
 * Re-designed Home Window